Project 3 Udacity/cd0039-Identity-and-Access-Management/Project/03_coffee_shop_full_stack/starter_code/backend/src/requirement.txt
alembic==1.7.7
aniso8601==6.0.0
asttokens==2.0.5
autopep8==1.6.0
Babel==2.9.0
backcall==0.2.0
black==22.3.0
cffi==1.15.1
click==8.1.3
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
export==0.2.0
filelock==3.7.0
Flask==2.1.2
Flask-Cors==3.0.7
Flask-Migrate==3.1.0
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==21.12.0
greenlet==1.1.2
idna==3.3
infinity==1.5
intervals==0.9.2
ipykernel==6.13.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.3.4
jupyter-core==4.10.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.5.2
postgres==4.0
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
pywin32==304
pyzmq==23.1.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.38.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.2.2.post1
validators==0.20.0
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.0.3
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
zope.event==4.5.0
zope.interface==5.4.0
